+++ druid/src/contexts.rs
@@ -16,7 +16,7 @@
 
 use std::{
     any::{Any, TypeId},
-    collections::VecDeque,
+    collections::{HashMap, VecDeque},
     ops::{Deref, DerefMut},
     rc::Rc,
     time::Duration,
@@ -61,6 +61,8 @@ pub(crate) struct ContextState<'a> {
     /// The id of the widget that currently has focus.
     pub(crate) focus_widget: Option<WidgetId>,
     pub(crate) root_app_data_type: TypeId,
+    pub(crate) timers: &'a mut HashMap<TimerToken, WidgetId>,
+    pub(crate) text_registrations: &'a mut Vec<TextFieldRegistration>,
 }
 
 /// A mutable context provided to event handling methods of widgets.
@@ -508,7 +510,7 @@ impl_context_method!(
         /// request with the event.
         pub fn request_timer(&mut self, deadline: Duration) -> TimerToken {
             trace!("request_timer deadline={:?}", deadline);
-            self.state.request_timer(&mut self.widget_state, deadline)
+            self.state.request_timer(self.widget_state.id, deadline)
         }
     }
 );
@@ -807,7 +809,7 @@ impl LifeCycleCtx<'_, '_> {
             document: Rc::new(document),
             widget_id: self.widget_id(),
         };
-        self.widget_state.text_registrations.push(registration);
+        self.state.text_registrations.push(registration);
     }
 
     /// Scrolls the area into view.
@@ -960,6 +962,8 @@ impl<'a> ContextState<'a> {
         window: &'a WindowHandle,
         window_id: WindowId,
         focus_widget: Option<WidgetId>,
+        timers: &'a mut HashMap<TimerToken, WidgetId>,
+        text_registrations: &'a mut Vec<TextFieldRegistration>,
     ) -> Self {
         ContextState {
             command_queue,
@@ -967,6 +971,8 @@ impl<'a> ContextState<'a> {
             window,
             window_id,
             focus_widget,
+            timers,
+            text_registrations,
             text: window.text(),
             root_app_data_type: TypeId::of::<T>(),
         }
@@ -978,10 +984,10 @@ impl<'a> ContextState<'a> {
             .push_back(command.default_to(self.window_id.into()));
     }
 
-    fn request_timer(&self, widget_state: &mut WidgetState, deadline: Duration) -> TimerToken {
+    fn request_timer(&mut self, widget_id: WidgetId, deadline: Duration) -> TimerToken {
         trace!("request_timer deadline={:?}", deadline);
         let timer_token = self.window.request_timer(deadline);
-        widget_state.add_timer(timer_token);
+        self.timers.insert(timer_token, widget_id);
         timer_token
     }
 }