+++ druid/src/core.rs
@@ -14,7 +14,7 @@
 
 //! The fundamental druid types.
 
-use std::collections::{HashMap, VecDeque};
+use std::collections::VecDeque;
 use tracing::{info_span, trace, warn};
 
 use crate::bloom::Bloom;
@@ -23,12 +23,10 @@ use crate::commands::SCROLL_TO_VIEW;
 use crate::contexts::ContextState;
 use crate::kurbo::{Affine, Insets, Point, Rect, Shape, Size, Vec2};
 use crate::sub_window::SubWindowUpdate;
-use crate::text::TextFieldRegistration;
-use crate::util::ExtendDrain;
 use crate::{
     ArcStr, BoxConstraints, Color, Command, Cursor, Data, Env, Event, EventCtx, InternalEvent,
     InternalLifeCycle, LayoutCtx, LifeCycle, LifeCycleCtx, Notification, PaintCtx, Region,
-    RenderContext, Target, TextLayout, TimerToken, UpdateCtx, Widget, WidgetId, WindowId,
+    RenderContext, Target, TextLayout, UpdateCtx, Widget, WidgetId, WindowId,
 };
 
 /// Our queue type
@@ -150,8 +148,6 @@ pub struct WidgetState {
     pub(crate) request_focus: Option<FocusChange>,
     pub(crate) children: Bloom<WidgetId>,
     pub(crate) children_changed: bool,
-    /// Associate timers with widgets that requested them.
-    pub(crate) timers: HashMap<TimerToken, WidgetId>,
     /// The cursor that was set using one of the context methods.
     pub(crate) cursor_change: CursorChange,
     /// The result of merging up children cursors. This gets cleared when merging state up (unlike
@@ -160,8 +156,6 @@ pub struct WidgetState {
 
     // Port -> Host
     pub(crate) sub_window_hosts: Vec<(WindowId, WidgetId)>,
-
-    pub(crate) text_registrations: Vec<TextFieldRegistration>,
 }
 
 /// Methods by which a widget can attempt to change focus state.
@@ -911,7 +905,9 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                     inner_ctx.is_handled = false;
                 } else if let Event::Notification(notification) = event {
                     // we will try again with the next parent
-                    inner_ctx.notifications.push_back(notification);
+                    inner_ctx
+                        .notifications
+                        .push_back(notification.with_route(self_id));
                 } else {
                     // could be unchecked but we avoid unsafe in druid :shrug:
                     unreachable!()
@@ -1252,12 +1248,10 @@ impl WidgetState {
             focus_chain: Vec::new(),
             children: Bloom::new(),
             children_changed: false,
-            timers: HashMap::new(),
             cursor_change: CursorChange::Default,
             cursor: None,
             sub_window_hosts: Vec::new(),
             is_explicitly_disabled_new: false,
-            text_registrations: Vec::new(),
             update_focus_chain: false,
         }
     }
@@ -1269,10 +1263,6 @@ impl WidgetState {
     pub(crate) fn tree_disabled_changed(&self) -> bool {
         self.children_disabled_changed
             || self.is_explicitly_disabled != self.is_explicitly_disabled_new
-    }
-
-    pub(crate) fn add_timer(&mut self, timer_token: TimerToken) {
-        self.timers.insert(timer_token, self.id);
     }
 
     /// Update to incorporate state changes from a child.
@@ -1314,9 +1304,6 @@ impl WidgetState {
         self.children_changed |= child_state.children_changed;
         self.request_update |= child_state.request_update;
         self.request_focus = child_state.request_focus.take().or(self.request_focus);
-        self.timers.extend_drain(&mut child_state.timers);
-        self.text_registrations
-            .append(&mut child_state.text_registrations);
         self.update_focus_chain |= child_state.update_focus_chain;
 
         // We reset `child_state.cursor` no matter what, so that on the every pass through the tree,
@@ -1384,8 +1371,9 @@ mod tests {
     use super::*;
     use crate::ext_event::ExtEventHost;
     use crate::text::ParseFormatter;
-    use crate::widget::{Flex, Scroll, Split, TextBox};
+    use crate::widget::{Button, Flex, Scroll, Split, TextBox};
     use crate::{WidgetExt, WindowHandle, WindowId};
+    use std::collections::HashMap;
     use test_env_log::test;
 
     const ID_1: WidgetId = WidgetId::reserved(0);
@@ -1424,12 +1412,16 @@ mod tests {
         let window = WindowHandle::default();
         let ext_host = ExtEventHost::default();
         let ext_handle = ext_host.make_sink();
+        let mut timers = Vec::new();
+        let mut text_registrations = HashMap::new();
         let mut state = ContextState::new::<Option<u32>>(
             &mut command_queue,
             &ext_handle,
             &window,
             WindowId::next(),
             None,
+            &mut text_registrations,
+            &mut timers,
         );
 
         let mut ctx = LifeCycleCtx {
@@ -1445,5 +1437,65 @@ mod tests {
         assert!(ctx.widget_state.children.may_contain(&ID_3));
         // A textbox is composed of three components with distinct ids
         assert_eq!(ctx.widget_state.children.entry_count(), 15);
+    }
+
+    #[test]
+    fn send_notifications() {
+        let mut widget = WidgetPod::new(Button::new("test".to_owned())).boxed();
+
+        let mut command_queue: CommandQueue = VecDeque::new();
+        let mut widget_state = WidgetState::new(WidgetId::next(), None);
+        let window = WindowHandle::default();
+        let ext_host = ExtEventHost::default();
+        let ext_handle = ext_host.make_sink();
+        let mut timers = Vec::new();
+        let mut text_registrations = HashMap::new();
+        let mut state = ContextState::new::<Option<u32>>(
+            &mut command_queue,
+            &ext_handle,
+            &window,
+            WindowId::next(),
+            None,
+            &mut text_registrations,
+            &mut timers,
+        );
+
+        let mut ctx = EventCtx {
+            widget_state: &mut widget_state,
+            notifications: &mut Default::default(),
+            is_handled: false,
+            state: &mut state,
+            is_root: false,
+        };
+
+        let ids = [
+            WidgetId::next(),
+            WidgetId::next(),
+            WidgetId::next(),
+            WidgetId::next(),
+        ];
+
+        let env = Env::with_default_i10n();
+
+        let notification = Command::new(druid::command::sys::CLOSE_WINDOW, (), Target::Global);
+
+        let mut notifictions = VecDeque::from(vec![
+            notification.clone().into_notification(ids[0]),
+            notification.clone().into_notification(ids[1]),
+            notification.clone().into_notification(ids[2]),
+            notification.into_notification(ids[3]),
+        ]);
+
+        widget.send_notifications(&mut ctx, &mut notifictions, &mut (), &env);
+
+        assert_eq!(ctx.notifications.len(), 4);
+        assert_eq!(ctx.notifications[0].source(), ids[0]);
+        assert_eq!(ctx.notifications[0].route(), widget.id());
+        assert_eq!(ctx.notifications[1].source(), ids[1]);
+        assert_eq!(ctx.notifications[1].route(), widget.id());
+        assert_eq!(ctx.notifications[2].source(), ids[2]);
+        assert_eq!(ctx.notifications[2].route(), widget.id());
+        assert_eq!(ctx.notifications[3].source(), ids[3]);
+        assert_eq!(ctx.notifications[3].route(), widget.id());
     }
 }